refactor(header): route header-no-glass through an index and drop the dead headerCloseControl export
The old headerCloseControl factory had no callers left but was still
exported, so a future screen picking it up would silently get the iOS 26
glass capsule back. header-no-glass also gains an index.ts like its
sibling components, replacing the deep imports.
