+++ app/screens/card-screen/card-dashboard-screen/card-dashboard-screen.tsx
@@ -15,7 +15,7 @@ import { NativeStackNavigationProp } from "@react-navigation/native-stack"
 import { BackendFeatureGate } from "@app/components/backend-feature-gate"
 import { GaloyIcon } from "@app/components/atomic/galoy-icon"
 import { BlinkCard } from "@app/components/blink-card"
-import { headerRightNoGlass } from "@app/components/header-no-glass/header-no-glass"
+import { headerRightNoGlass } from "@app/components/header-no-glass"
 import {
   CardActionButtons,
   CardBalanceSection,
